+++ kitupiikki/raportti/paivakirjaraportti.h
@@ -53,7 +53,7 @@ public:
                                  bool ryhmitalajeittain = false, bool tulostakohdennukset = false,
                                  bool tulostasummat = false);
 protected:
-    static void kirjoitaSummaRivi(RaportinKirjoittaja &rk, int debet, int kredit, int sarakeleveys);
+    static void kirjoitaSummaRivi(RaportinKirjoittaja &rk, qlonglong debet, qlonglong kredit, int sarakeleveys);
 
     Ui::Paivakirja *ui;
 };
